+++ matchinggame/css/api.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title></title>
+    <link rel="stylesheet" href="/css/style.css">
+  </head>
+  <body>
+    <h1>API Loading</h1>
+    <script type="text/javascript" src="/js/main.js"></script>
+  </body>
+</html>
