+++ matchinggame/css/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "figlet": "^1.8.0"
+  }
+}
